chore(cordyceps): put list tests in their own file (#203)
